Support Marshal interface for serializization
Marshal refuses to serialize NBayes::Base because NBayes::Data contains
a hash with a default.  This implements marshal_dump and marshal_load to
allow it.  It also uses the opportunity to group the serialized data by
frequency instead of token, reducing serialized data size by around a third
with trained classifiers.